Kanban UI

+++ components/dialog/allProjectsDialog.tsx
@@ -13,6 +13,22 @@ import { Input } from "@/components/ui/input";
 import { CreateProjectDialog } from "@/components/dialog/createProjectDialog";
 import { TaskCount } from "@/components/cards/taskCount";
 import { EditProject } from "@/components/popover/editProject";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipTrigger,
+} from "@/components/ui/tooltip";
 
 import { Project } from "@/types/project/projectType";
 
@@ -25,29 +41,31 @@ export function AllProjectsDialog({
   projectId: string;
   projects: Project[];
 }) {
-  const { deleteProject } = useProject();
+  const { deleteProject, deleteAllProjects } = useProject();
 
+  // Add Error | Success message
   const handleDeleteProject = (id: string) => deleteProject(id);
+  const handleDeleteAllProjects = () => deleteAllProjects();
 
   return (
     <Dialog>
       <DialogTrigger asChild>
-        <SquarePlus />
+        <SquarePlus className="lg:size-6" />
       </DialogTrigger>
       <DialogContent className=" sm:max-w-[425px]">
         <DialogHeader className="flex fle-col">
           <DialogTitle>All Projects</DialogTitle>
           <Separator className="my-3" />
-          <div className="flex justify-between gap-4">
+          <div className="flex flex-col lg:flex-row justify-between gap-4">
             <Input
               type="text"
-              className="w-1/2"
+              className="lg:w-1/2"
               placeholder="Search your project"
             />
             <CreateProjectDialog />
           </div>
         </DialogHeader>
-        <div className="flex flex-col gap-3">
+        <div className="flex flex-col gap-3 truncate">
           {projects.map((project) => (
             <div
               key={project.id}
@@ -56,9 +74,17 @@ export function AllProjectsDialog({
               <div className="bg-priority-low p-1 rounded-lg">
                 <SquareKanban className="text-primary-foreground" />
               </div>
-              <div>
+              <div className="truncate">
                 <h3 className="font-bold flex gap-2 items-center">
-                  {project.title}
+                  <Tooltip>
+                    <TooltipTrigger asChild>
+                      <h5 className="truncate">{project.title}</h5>
+                    </TooltipTrigger>
+                    <TooltipContent className="text-lg text-foreground">
+                      <h5>{project.title}</h5>
+                    </TooltipContent>
+                  </Tooltip>
+
                   {project.id === projectId && (
                     <span className="text-sm bg-muted-foreground text-muted p-1.5 rounded-2xl">
                       Select?
@@ -72,22 +98,69 @@ export function AllProjectsDialog({
               </div>
               <div className="flex flex-1 justify-end gap-3">
                 <EditProject project={project} />
-                <Button
-                  variant="ghost"
-                  className="cursor-pointer"
-                  onClick={() => handleDeleteProject(project.id)}
-                >
-                  <Trash className="text-destructive" />
-                </Button>
+                <AlertDialog>
+                  <AlertDialogTrigger asChild>
+                    <Button variant="ghost" className="cursor-pointer">
+                      <Trash className="text-destructive" />
+                    </Button>
+                  </AlertDialogTrigger>
+                  <AlertDialogContent>
+                    <AlertDialogHeader>
+                      <AlertDialogTitle>Delete This Project?</AlertDialogTitle>
+                      <AlertDialogDescription>
+                        This action cannot be undone. This will permanently
+                        delete this project and tasks from our servers.
+                      </AlertDialogDescription>
+                    </AlertDialogHeader>
+                    <AlertDialogFooter>
+                      <AlertDialogCancel className="cursor-pointer">
+                        Cancel
+                      </AlertDialogCancel>
+                      <AlertDialogAction
+                        onClick={() => handleDeleteProject(project.id)}
+                        className="bg-destructive hover:bg-destructive/80 cursor-pointer"
+                      >
+                        Delete
+                      </AlertDialogAction>
+                    </AlertDialogFooter>
+                  </AlertDialogContent>
+                </AlertDialog>
               </div>
             </div>
           ))}
         </div>
         <DialogFooter className="flex !justify-between items-center">
           <span className="text-sm">{projects.length} Projects</span>
-          <Button variant="destructive" className="cursor-pointer">
-            Delete all
-          </Button>
+          <AlertDialog>
+            <AlertDialogTrigger asChild>
+              <Button
+                variant="destructive"
+                className="cursor-pointer w-full lg:w-auto"
+              >
+                Delete all
+              </Button>
+            </AlertDialogTrigger>
+            <AlertDialogContent>
+              <AlertDialogHeader>
+                <AlertDialogTitle>Delete All Projects?</AlertDialogTitle>
+                <AlertDialogDescription>
+                  This action cannot be undone. This will permanently delete all
+                  your projects and tasks from our servers.
+                </AlertDialogDescription>
+              </AlertDialogHeader>
+              <AlertDialogFooter>
+                <AlertDialogCancel className="cursor-pointer">
+                  Cancel
+                </AlertDialogCancel>
+                <AlertDialogAction
+                  onClick={handleDeleteAllProjects}
+                  className="bg-destructive hover:bg-destructive/80 cursor-pointer"
+                >
+                  Delete
+                </AlertDialogAction>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
         </DialogFooter>
       </DialogContent>
     </Dialog>
